--- middleware/cache.js
@@ -1,3 +1,0 @@
-export const getDataCache = (req, res) => {
-  return res.response({ message: `Returned from cache layer` });
-};
